sentinel_core/mcp/requirements.txt
fastapi
uvicorn
pydantic-settings
redis
pyjwt
cryptography
pydantic
